fixed some of the godoc tests
