+++ scheduler/indexer/godoc/interfaces.go
@@ -10,9 +10,11 @@ type PackageMetadata struct {
 	repo        string
 }
 
-// FetchPackageMetadataArgs lol
+// FetchPackageMetadataArgs is the args struct for fetching package metadata
+// from godoc.
 type FetchPackageMetadataArgs struct {
 	ParseHTML htmlParser
 }
 
+// htmlParser parses an HTML doc from a url and returns a goquery doc.
 type htmlParser func(url string) (*goquery.Document, error)
